+++ CHANGELOG.md
@@ -59,6 +59,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2252 Bump auth0-react from 1.9.0 to 2.0.1
 - YLP-2256 Bump minor deps
 - YLP-2258 Bump react-markdown
+- YLP-2289 Migrate Standard Deviation from stat.js to yourloops
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team